tensor.result_type now also handles Python scalars
It handles scalars consistently with the way elementwise
operations do, following NEP-0050.

Some definitions were migrated from _elementwise_common
to _type_utils.
